backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
